+++ JavaGWT.RayTracerDemo/src/main/java/io/github/timeu/javagwtraytracerdemo/shared/RTObject.java
@@ -6,7 +6,7 @@ package io.github.timeu.javagwtraytracerdemo.shared;
 public abstract class RTObject {
     public Color color;
 
-    public abstract float Intersect(Ray ray);
+    public abstract double Intersect(Ray ray);
 
     public abstract Vector3f GetSurfaceNormalAtPoint(Vector3f p);
